Converted pydantic things to @dataclass definitions.
